+++ at_follows_flutter/lib/services/sdk_service.dart
@@ -29,6 +29,8 @@ class SDKService {
     this._atsign = _atClientServiceInstance.atClient!.currentAtSign;
     Strings.rootdomain =
         _atClientServiceInstance.atClient!.preference!.rootDomain;
+    AppConstants.appNamespace =
+        _atClientServiceInstance.atClient!.preference!.namespace!;
   }
 
   get atsign => this._atsign;
@@ -64,7 +66,8 @@ class SDKService {
       fromDate = fromDate.split(' ')[0];
     }
     var response = await _atClientServiceInstance.atClient!.notifyList(
-        regex: ('${AppConstants.following}|${AppConstants.followers}'),
+        regex:
+            ('${AppConstants.containsFollowing}|${AppConstants.containsFollowers}'),
         fromDate: fromDate);
     response = response.toString().replaceAll('data:', '');
     if (response == 'null') {
@@ -118,8 +121,27 @@ class SDKService {
         .getAtKeys(regex: regex)
         .timeout(Duration(seconds: AppConstants.responseTimeLimit),
             onTimeout: () => _onTimeOut());
-    AtFollowsValue value =
-        scanKey.isNotEmpty ? await this.get(scanKey[0]) : AtFollowsValue();
+    AtFollowsValue value = scanKey.isNotEmpty
+        ? await this.get(_formKeysFromScanKey(scanKey[0]))
+        : AtFollowsValue();
+    //migrates to newnamespace
+    if (scanKey.isNotEmpty &&
+        _isOldKey(scanKey[0].key!) &&
+        value.value != null) {
+      var metadata = scanKey[0].metadata;
+      metadata!.namespaceAware = false;
+      var newKey = AtKey()..metadata = metadata;
+      newKey.key = scanKey[0].key!.contains('following')
+          ? AppConstants.followingKey
+          : AppConstants.followersKey;
+      await this.put(newKey, value.value);
+      value = await this.get(newKey);
+      if (value.value != null) {
+        scanKey[0].metadata!.namespaceAware = true;
+        await this.delete(scanKey[0]);
+      }
+    }
+
     return value;
   }
 
@@ -163,5 +185,18 @@ class SDKService {
     response = response.toString().replaceAll(RegExp('[|]'), '');
     response = response.replaceAll('notification:', '').trim();
     return AtNotification.fromJson(jsonDecode(response));
+  }
+
+  ///Returns `true` if key is old key else `false`.
+  bool _isOldKey(String key) {
+    return !key.contains(AppConstants.libraryNamespace);
+  }
+
+  //Returns AtKey by formatting scanned [atKey].
+  AtKey _formKeysFromScanKey(AtKey atkey) {
+    var atKey = atkey;
+    atKey.key = atKey.key! + '.' + atKey.namespace!;
+    atKey.metadata!.namespaceAware = false;
+    return atKey;
   }
 }