+++ at_follows_flutter/lib/utils/strings.dart
@@ -5,6 +5,7 @@ class Strings {
   static const String Following = 'Following';
   static const String Follow = 'Follow';
   static const String Unfollow = 'Unfollow';
+  static const String Remove = 'Remove';
   static const String BackButton = 'Back';
   static const String Title = 'Connections';
   static const String Search = 'Filter @signs';
@@ -71,6 +72,11 @@ class Strings {
   static const String closeButton = 'Close';
 
   //loading texts
-  static const String loadingDescription =
-      'Please wait while loading your data';
+  static const String getAtSignList =
+      'Please wait and don\'t go back while fetching your data';
+  static const String statusCheck = 'Fetching @sign status...';
+  static const String followMessage =
+      'Please wait while adding @sign to your following list';
+  static const String unfollowMessage =
+      'Please wait while removing @sign from list';
 }