+++ at_follows_flutter/lib/widgets/followers.dart
@@ -62,6 +62,12 @@ class _FollowersState extends State<Followers> {
                   ),
                 ),
                 SizedBox(height: 5.0.toHeight),
+                if (provider.message != null)
+                  Text(
+                    provider.message!,
+                    style: CustomTextStyles.fontBold14primary,
+                    textAlign: TextAlign.center,
+                  )
               ],
             ),
           ),
@@ -198,45 +204,71 @@ class _FollowersState extends State<Followers> {
                                             color: Colors.black,
                                             size: 25.toFont),
                                   ),
-                                  title: Text(currentAtsign.title!,
-                                      style: CustomTextStyles.fontR16primary),
+                                  title: Row(
+                                    mainAxisSize: MainAxisSize.min,
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.center,
+                                    // mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      Flexible(
+                                        // flex: 3,
+                                        // fit: FlexFit.loose,
+                                        child: Text((currentAtsign.title!),
+                                            overflow: TextOverflow.ellipsis,
+                                            softWrap: false,
+                                            style: CustomTextStyles
+                                                .fontR16primary),
+                                      ),
+                                      Text(' \u2022 ',
+                                          style: TextStyle(
+                                            fontSize: 25.toFont,
+                                            fontWeight: FontWeight.bold,
+                                          )),
+                                      currentAtsign.isFollowing!
+                                          ? GestureDetector(
+                                              onTap: () async {
+                                                widget.isFollowing
+                                                    ? await provider.unfollow(
+                                                        currentAtsign.title)
+                                                    : await provider.delete(
+                                                        currentAtsign.title!);
+                                                sortedListWithAlphabet[index]
+                                                    .isFollowing = false;
+                                                setState(() {});
+                                              },
+                                              child: Text(
+                                                  widget.isFollowing
+                                                      ? Strings.Following
+                                                      : Strings.Remove,
+                                                  style: TextStyle(
+                                                      fontSize: 14.toFont,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                      color: Colors.grey)),
+                                            )
+                                          : GestureDetector(
+                                              onTap: () async {
+                                                provider.follow(
+                                                    currentAtsign.title);
+                                                sortedListWithAlphabet[index]
+                                                    .isFollowing = true;
+                                                setState(() {});
+                                              },
+                                              child: Text(Strings.Follow,
+                                                  style: TextStyle(
+                                                      fontSize: 14.toFont,
+                                                      fontWeight:
+                                                          FontWeight.bold,
+                                                      color: ColorConstants
+                                                          .activeColor)),
+                                            )
+                                    ],
+                                  ),
                                   subtitle: currentAtsign.subtitle != null
                                       ? Text(currentAtsign.subtitle!,
                                           style:
                                               CustomTextStyles.fontR14primary)
                                       : null,
-                                  trailing: Row(
-                                    mainAxisSize: MainAxisSize.min,
-                                    children: [
-                                      CustomButton(
-                                          height: 35.toHeight,
-                                          // width: 70.toWidth,
-                                          // width: 150.toWidth,
-                                          isActive: !currentAtsign.isFollowing!,
-                                          onPressedCallBack: (value) async {
-                                            if (value) {
-                                              provider.unfollow(
-                                                  currentAtsign.title);
-                                            } else {
-                                              provider
-                                                  .follow(currentAtsign.title);
-                                            }
-                                            sortedListWithAlphabet[index]
-                                                .isFollowing = !value;
-                                            setState(() {});
-                                          },
-                                          text: currentAtsign.isFollowing!
-                                              ? Strings.Unfollow
-                                              : Strings.Follow),
-                                      // IconButton(
-                                      //     iconSize: 20.toFont,
-                                      //     icon: Icon(Icons.delete),
-                                      //     onPressed: () async {
-                                      //       await provider
-                                      //           .delete(currentAtsign.title);
-                                      //     }),
-                                    ],
-                                  ),
                                 ),
                               );
                             },
@@ -254,6 +286,7 @@ class _FollowersState extends State<Followers> {
           ),
         );
       } else if (provider.status == Status.error) {
+        _logger.severe('Loading throws ${provider.error}');
         return AtExceptionHandler().handle(provider.error, context);
       } else {
         WidgetsBinding.instance!.addPostFrameCallback((_) async {